Add ability to store and get accounts from database

+++ cps/database/schemas/smartgrow.sql
@@ -21,6 +21,6 @@ CREATE TABLE plant_data(
 );
 
 CREATE TABLE leaf_accounts(
-    leaf_id NUMERIC PRIMARY KEY,
+    leaf_id SERIAL PRIMARY KEY,
     macAddress TEXT
 );
